pyproject.toml
[tool.poetry]
name = "mudi"
version = "0.1.0"
description = "An opinionated static site generator without bells and whistles."
authors = ["Dan Turkel <daturkel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
python-frontmatter = "^0.5.0"
marko = {extras = ["codehilite", "footnote"], version = "^0.8.2"}
pydantic = "^1.5.1"
jinja2 = "^2.11.2"
libsass = "^0.20.0"
toml = "^0.10.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
